# api/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QRIS Payment API",
    "description": "QRIS Payment Processing API with HMAC-SHA256 Authentication, Redis Caching, and PostgreSQL",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/qris/inquiry/{qris_payload}": {
      "get": {
        "summary": "QRIS Inquiry",
        "description": "Parse QRIS payload and return merchant information. Results are cached in Redis.",
        "tags": [
          "QRIS"
        ],
        "parameters": [
          {
            "name": "qris_payload",
            "in": "path",
            "required": true,
            "description": "QRIS EMV QR payload string",
            "schema": {
              "type": "string"
            },
            "example": "00020101021126690021ID.CO.BANKMANDIRI.WWW01189360000801299399930211712993999340303UKE51440014ID.CO.QRIS.WWW0215ID10232756067300303UKE5204274153033605802ID5912M Ivan Store6015Jakarta Timur"
          },
          {
            "name": "X-Client-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "MK-9921-X"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2026-02-25T20:30:00Z"
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "description": "HMAC-SHA256(SecretKey, Method+Path+Timestamp+Body)",
            "schema": {
              "type": "string"
            },
            "example": "a5f8e..."
          }
        ],
        "responses": {
          "200": {
            "description": "Inquiry successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Merchant not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/qris/payment": {
      "post": {
        "summary": "QRIS Payment",
        "description": "Process a QRIS payment. Validates inquiry ID, verifies PIN, deducts balance with optimistic locking.",
        "tags": [
          "QRIS"
        ],
        "parameters": [
          {
            "name": "X-Client-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "MK-9921-X"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2026-02-25T20:30:00Z"
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a5f8e..."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/transaction/status/{transaction_id}": {
      "get": {
        "summary": "Get Transaction Status",
        "description": "Retrieve the current status of a transaction by its UUID.",
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "description": "Transaction UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          {
            "name": "X-Client-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "MK-9921-X"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2026-02-25T20:30:00Z"
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a5f8e..."
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStatusApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InquiryData": {
        "type": "object",
        "properties": {
          "merchant_id": {
            "type": "string",
            "example": "MICH-001"
          },
          "merchant_name": {
            "type": "string",
            "example": "Toko Berkah Mandiri"
          },
          "terminal_id": {
            "type": "string",
            "example": "T001"
          },
          "city": {
            "type": "string",
            "example": "Malang"
          },
          "fixed_amount": {
            "type": "number",
            "example": 0
          },
          "inquiry_id": {
            "type": "string",
            "example": "inq_789abc"
          }
        }
      },
      "Metadata": {
        "type": "object",
        "properties": {
          "latency_ms": {
            "type": "integer",
            "example": 45
          },
          "source": {
            "type": "string",
            "enum": [
              "cache",
              "database"
            ],
            "example": "cache"
          }
        }
      },
      "InquiryApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/InquiryData"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        }
      },
      "PaymentRequest": {
        "type": "object",
        "required": [
          "inquiry_id",
          "user_id",
          "amount",
          "payment_method",
          "pincode"
        ],
        "properties": {
          "inquiry_id": {
            "type": "string",
            "example": "inq_789abc"
          },
          "user_id": {
            "type": "string",
            "example": "user_123"
          },
          "amount": {
            "type": "number",
            "example": 50000
          },
          "payment_method": {
            "type": "string",
            "example": "balance"
          },
          "pincode": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "PaymentData": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "processing"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "message": {
            "type": "string",
            "example": "Transaksi sedang diproses"
          },
          "estimated_completion": {
            "type": "string",
            "example": "200ms"
          }
        }
      },
      "PaymentApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PaymentData"
          }
        }
      },
      "TransactionStatusData": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED",
              "REVERSED"
            ],
            "example": "SUCCESS"
          },
          "final_balance": {
            "type": "number",
            "example": 450000
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-25T20:15:30Z"
          }
        }
      },
      "TransactionStatusApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TransactionStatusData"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "errors": {
            "type": "string",
            "example": "Invalid signature"
          }
        }
      }
    }
  }
}